Create google image search automated.json
will take an image you specify and search it on google images.  once image is found or similiar image is found or keyword in the url links of the images on the results page it will click on whatever you specify for it to click on such as the first image with a link to ebay.